requirements.txt
pytest>=7.4.0
selenium>=4.15.0
webdriver-manager>=4.0.0
Appium-Python-Client>=3.1.0
